requirements.txt
pandas~=2.2.0
flask~=3.0.2
pytils~=0.4.1
pytz~=2024.1
python-dateutil~=2.8.2